Chaning a unit test to integration test
