+++ tests/integration/gcp/conftest.py
@@ -1,6 +1,7 @@
 import pytest
 import random
 import fairing
+import httplib2
 from google.cloud import storage
 
 GCS_PROJECT_ID = fairing.cloud.gcp.guess_project_name()
@@ -16,3 +17,21 @@ def temp_gcs_prefix():
     for blob in bucket.list_blobs(prefix=rnd_prefix):
         print("deleting {}".format(blob.name))
         blob.delete()
+
+@pytest.fixture
+def httpmock():
+    http = httplib2.Http()
+    class HTTPMock:
+
+        def __init__(self, *args, **kwargs):
+            pass
+                     
+        def request(self, *args, **kwargs):
+            if args and len(args)>=4:
+                headers = args[3] or {}
+            else:
+                headers = kwargs.get('headers') or {}
+            print("HTTPMock url:{} user-agent:{}".format(args[0], headers.get('user-agent')))
+            return http.request(*args, **kwargs)
+    
+    return HTTPMock
